Damn thing refuses to post, cant even test the nodemailer part errgegrgergqijewbfkhbqr;lkgjq'elrkj
